feat: rename Entity\Date* traits
