+++ src/Entity/Gateway/Charge.php
@@ -4,10 +4,10 @@ namespace App\Entity\Gateway;
 
 use App\Entity\Accounting\Accounting;
 use App\Entity\Accounting\Transaction;
+use App\Entity\DateCreatedTrait;
+use App\Entity\DateUpdatedTrait;
 use App\Entity\EmbeddableMoney as Money;
 use App\Entity\MigratedTrait;
-use App\Entity\Trait\TimestampedCreationEntity;
-use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Gateway\ChargeStatus;
 use App\Gateway\ChargeType;
 use App\Mapping\Provider\EntityMapProvider;
@@ -31,8 +31,8 @@ use Symfony\Component\Validator\Constraints as Assert;
 class Charge
 {
     use MigratedTrait;
-    use TimestampedCreationEntity;
-    use TimestampedUpdationEntity;
+    use DateCreatedTrait;
+    use DateUpdatedTrait;
 
     #[ORM\Id]
     #[ORM\GeneratedValue]
