+++ src/Entity/Project/Collaboration.php
@@ -2,10 +2,10 @@
 
 namespace App\Entity\Project;
 
+use App\Entity\DateCreatedTrait;
+use App\Entity\DateUpdatedTrait;
 use App\Entity\LocalizedInterface;
 use App\Entity\LocalizedTrait;
-use App\Entity\Trait\TimestampedCreationEntity;
-use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\CollaborationRepository;
 use AutoMapper\Attribute\MapProvider;
@@ -19,8 +19,8 @@ use Gedmo\Mapping\Annotation as Gedmo;
 class Collaboration implements LocalizedInterface
 {
     use LocalizedTrait;
-    use TimestampedCreationEntity;
-    use TimestampedUpdationEntity;
+    use DateCreatedTrait;
+    use DateUpdatedTrait;
 
     #[ORM\Id]
     #[ORM\GeneratedValue]
