+++ src/Entity/Project/Update.php
@@ -2,10 +2,10 @@
 
 namespace App\Entity\Project;
 
+use App\Entity\DateCreatedTrait;
+use App\Entity\DateUpdatedTrait;
 use App\Entity\LocalizedInterface;
 use App\Entity\LocalizedTrait;
-use App\Entity\Trait\TimestampedCreationEntity;
-use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Entity\User\User;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\UpdateRepository;
@@ -20,8 +20,8 @@ use Gedmo\Mapping\Annotation as Gedmo;
 class Update implements LocalizedInterface
 {
     use LocalizedTrait;
-    use TimestampedCreationEntity;
-    use TimestampedUpdationEntity;
+    use DateCreatedTrait;
+    use DateUpdatedTrait;
 
     #[ORM\Id]
     #[ORM\GeneratedValue]
