+++ src/Entity/User/User.php
@@ -3,12 +3,12 @@
 namespace App\Entity\User;
 
 use App\Entity\Accounting\Accounting;
+use App\Entity\DateCreatedTrait;
+use App\Entity\DateUpdatedTrait;
 use App\Entity\DedupedTrait;
 use App\Entity\Interface\AccountingOwnerInterface;
 use App\Entity\MigratedTrait;
 use App\Entity\Project\Project;
-use App\Entity\Trait\TimestampedCreationEntity;
-use App\Entity\Trait\TimestampedUpdationEntity;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\User\UserRepository;
 use AutoMapper\Attribute\MapProvider;
@@ -37,8 +37,8 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, Account
 {
     use MigratedTrait;
     use DedupedTrait;
-    use TimestampedCreationEntity;
-    use TimestampedUpdationEntity;
+    use DateCreatedTrait;
+    use DateUpdatedTrait;
 
     #[ORM\Id]
     #[ORM\GeneratedValue]
