+++ src/Entity/User/UserToken.php
@@ -2,8 +2,8 @@
 
 namespace App\Entity\User;
 
+use App\Entity\DateCreatedTrait;
 use App\Entity\Interface\UserOwnedInterface;
-use App\Entity\Trait\TimestampedCreationEntity;
 use App\Entity\Trait\UserOwnedTrait;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\User\UserTokenRepository;
@@ -23,7 +23,7 @@ use Doctrine\ORM\Mapping as ORM;
 #[ORM\Entity(repositoryClass: UserTokenRepository::class)]
 class UserToken implements UserOwnedInterface
 {
-    use TimestampedCreationEntity;
+    use DateCreatedTrait;
     use UserOwnedTrait;
 
     #[ORM\Id]
